moved utils to folder utils
to avoid name conflicts
